+++ api/test/service/mocks_test.go
@@ -97,7 +97,7 @@ func (m *MockCardRepository) UpdatePrice(id uint, newPrice float32) (*model.Card
 	return nil, args.Error(1)
 }
 
-func (m *MockCardRepository) UpdateInStockAmount(id uint, newPrice float32) (*model.Card, error) {
+func (m *MockCardRepository) UpdateInStockAmount(id uint, newPrice uint) (*model.Card, error) {
 	args := m.Called(id, newPrice)
 	switch card := args.Get(0).(type) {
 	case *model.Card:
